requirements.txt
asttokens==2.4.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
libretranslatepy==2.1.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
rich==13.6.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
textblob==0.17.1
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
translate==3.6.1
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
